FAT2-212 remove garbage fields from assessor response.
